afubot/bot/requirements.txt
python-telegram-bot
python-dotenv
requests
PyMySQL>=1.1.0
